requirements.txt
fastapi
uvicorn
scikit-learn
pandas
gdown
joblib
tensorflow
Pillow
python-multipart
pydantic
aiohttp
python-dotenv
pytest
loguru
